chore: 데이터 모델 분리, 일부 데이터 형 변경
