--- OpenMarket/OpenMarket/Models.swift
@@ -1,123 +1,0 @@
-import Foundation
-import CoreText
-
-enum Currency: Decodable {
-    case KRW
-    case USD
-}
-
-struct Product: Decodable {
-    let id: Int
-    let vendorId: Int
-    let name: String
-    let thumbnail: String
-    let currency: String
-    let price: Double
-    let bargainPrice: Double
-    let discountedPrice: Double
-    let stock: Int
-    let createdAt: String
-    let issuedAt: String
-    
-    private enum CodingKeys: String, CodingKey {
-        case id
-        case vendorId = "vendor_id"
-        case name
-        case thumbnail
-        case currency
-        case price
-        case bargainPrice = "bargain_price"
-        case discountedPrice = "discounted_price"
-        case stock
-        case createdAt = "created_at"
-        case issuedAt = "issued_at"
-    }
-}
-
-struct ProductList: Decodable {
-    let pageNo: Int
-    let itemsPerPage: Int
-    let totalCount: Int
-    let offset: Int
-    let limit: Int
-    let lastPage: Int
-    let hasNext: Bool
-    let hasPrev: Bool
-    let products: [Product]
-    
-    private enum CodingKeys: String, CodingKey {
-        case pageNo = "page_no"
-        case itemsPerPage = "items_per_page"
-        case totalCount = "total_count"
-        case offset
-        case limit
-        case lastPage = "last_page"
-        case hasNext = "has_next"
-        case hasPrev = "has_prev"
-        case products = "pages"
-    }
-}
-
-struct ProductDetailImage: Decodable {
-    let id: Int
-    let url: String
-    let thumbnailUrl: String
-    let uploadResult: Bool
-    let issuedAt: String
-    
-    private enum CodingKeys: String, CodingKey {
-        case id
-        case url
-        case thumbnailUrl = "thumbnail_url"
-        case uploadResult = "succeed"
-        case issuedAt = "issued_at"
-    }
-}
-
-struct Vendor: Decodable {
-    let name: String
-    let id: Int
-    let createdAt: String
-    let issuedAt: String
-    
-    private enum CodingKeys: String, CodingKey {
-        case name
-        case id
-        case createdAt = "created_at"
-        case issuedAt = "issued_at"
-    }
-}
-
-struct ProductDetail: Decodable {
-    let id: Int
-    let vendorId: Int
-    let name: String
-    let description: String
-    let thumbnail: String
-    let currency: String
-    let price: Double
-    let bargainPrice: Double
-    let discountedPrice: Double
-    let stock: Int
-    let images: [ProductDetailImage]
-    let vendors: Vendor
-    let createdAt: String
-    let issuedAt: String
-    
-    private enum CodingKeys: String, CodingKey {
-        case id
-        case vendorId = "vendor_id"
-        case name
-        case description
-        case thumbnail
-        case currency
-        case price
-        case bargainPrice = "bargain_price"
-        case discountedPrice = "discounted_price"
-        case stock
-        case images
-        case vendors
-        case createdAt = "created_at"
-        case issuedAt = "issued_at"
-    }
-}
