+++ OpenMarket/OpenMarket/Models/Product.swift
@@ -1,0 +1,27 @@
+struct Product: Decodable {
+    let id: Int
+    let vendorId: Int
+    let name: String
+    let thumbnail: String
+    let currency: String
+    let price: Double
+    let bargainPrice: Double
+    let discountedPrice: Double
+    let stock: Int
+    let createdAt: String
+    let issuedAt: String
+    
+    private enum CodingKeys: String, CodingKey {
+        case id
+        case vendorId = "vendor_id"
+        case name
+        case thumbnail
+        case currency
+        case price
+        case bargainPrice = "bargain_price"
+        case discountedPrice = "discounted_price"
+        case stock
+        case createdAt = "created_at"
+        case issuedAt = "issued_at"
+    }
+}
