+++ OpenMarket/OpenMarket/Models/ProductDetail.swift
@@ -1,0 +1,35 @@
+import Foundation
+
+struct ProductDetail: Decodable {
+    let id: Int
+    let vendorId: Int
+    let name: String
+    let description: String
+    let thumbnail: String
+    let currency: String
+    let price: Double
+    let bargainPrice: Double
+    let discountedPrice: Double
+    let stock: Int
+    let images: [ProductDetailImage]
+    let vendors: Vendor
+    let createdAt: Date
+    let issuedAt: Date
+    
+    private enum CodingKeys: String, CodingKey {
+        case id
+        case vendorId = "vendor_id"
+        case name
+        case description
+        case thumbnail
+        case currency
+        case price
+        case bargainPrice = "bargain_price"
+        case discountedPrice = "discounted_price"
+        case stock
+        case images
+        case vendors
+        case createdAt = "created_at"
+        case issuedAt = "issued_at"
+    }
+}
